refactor: 시간 데이터 custom hook 분리
