+++ src/hooks/home/useFormData.tsx
@@ -10,14 +10,11 @@ const useFormData = (data: FormDataType) => {
 
     for (const key in data) {
       if (data[key] !== null) {
-        // 'board_img' 키일 경우 확인
         if (key === 'board_img') {
-          // board_img가 null이 아닌 경우에만 formData에 추가
           if (data.board_img) {
             formData.append(key, data.board_img);
           }
         } else {
-          // 다른 키들은 null이 아니면 formData에 추가
           formData.append(key, `${data[key]}`);
         }
       }
